feat: introduce new parameter `--ignore-password`
The bot was writing Shopee username and password into cookies file in
plain text. This is because when the cookies expire, the bot can
re-login by username and password, and this process may not require
further authentication if the bot is using the same IP and browser.

However, saving plain text password in a file raises the risk, so we add
an option to let user not to save password in cookies file.