docs: deprecate parameter `--ignore-password` in README [no ci]
Parameter `--ignore-password` is deprecated as the bot will not work
anymore when the cookies expire. The bot is designed to be run
periodically without human interaction, and thus it should has the
capability to re-login to Shopee using username and password. Therefore,
we now force users to put username and password in the credential file,
and it is user's responsibility to keep that file safe.